refactor(babylon): moving input detection to a dedicated helper class
